requirements.txt
# FastAPI 相关
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=2.0
pydantic-settings>=2.0

# 数据处理
numpy
pandas
scipy

# 深度学习
torch
torchvision
transformers

# 视频处理
opencv-python

# 后台任务
celery
redis

# 面部检测
batch-face>=1.5.0
py-feat>=0.6.2

# 工具
pyyaml
python-dotenv
python-multipart

# 其他
requests<3.0.0
